refactor: move determine_energy_source utility function to save/utils.py module
